+++ src/compiler/ast.hpp
@@ -6,6 +6,7 @@
 #include <list>
 #include <vector>
 #include <utility>
+#include <optional>
 
 namespace anole
 {
@@ -18,10 +19,7 @@ using ExprList
     = std::list<Ptr<struct Expr>>
 ;
 using DeclList
-    = std::list<Ptr<struct VariableDeclarationStmt>>
-;
-using IdentList
-    = std::list<Ptr<struct IdentifierExpr>>
+    = std::list<Ptr<struct DeclarationStmt>>
 ;
 using ArgumentList
     = std::list<std::pair<Ptr<struct Expr>, bool>>
@@ -196,12 +194,11 @@ struct LambdaExpr : Expr
 struct DotExpr : Expr
 {
     Ptr<Expr> left;
-    Ptr<IdentifierExpr> id;
+    std::string name;
 
-    DotExpr(Ptr<Expr> &&left,
-        Ptr<IdentifierExpr> &&id)
+    DotExpr(Ptr<Expr> &&left, std::string name)
       : left(std::move(left))
-      , id(std::move(id))
+      , name(std::move(name))
     {
         // ...
     }
@@ -211,7 +208,7 @@ struct DotExpr : Expr
 
 struct EnumExpr : Expr
 {
-    DeclList decls;
+    std::list<struct VariableDeclarationStmt> decls;
 
     EnumExpr() noexcept = default;
 
@@ -263,6 +260,25 @@ struct DictExpr : Expr
     void codegen(Code &) override;
 };
 
+struct ClassExpr : Expr
+{
+    String name;
+    ArgumentList bases;
+    DeclList members;
+
+    ClassExpr(String name,
+        ArgumentList &&bases,
+        DeclList &&members)
+      : name(std::move(name))
+      , bases(std::move(bases))
+      , members(std::move(members))
+    {
+        // ...
+    }
+
+    void codegen(Code &) override;
+};
+
 struct DelayExpr : Expr
 {
     Ptr<Expr> expr;
@@ -307,8 +323,8 @@ struct UseStmt : Stmt
         Type type;
     };
 
-    using Alias = std::pair<Module, String>;
     // second String is the alias
+    using Alias = std::pair<Module, String>;
     using Aliases = std::list<Alias>;
 
     // aliases are empty means `use *`
@@ -340,17 +356,21 @@ struct ExprStmt : Stmt
     void codegen(Code &) override;
 };
 
-struct VariableDeclarationStmt : Stmt
+struct DeclarationStmt : Stmt
 {
-    Ptr<IdentifierExpr> id;
+    virtual ~DeclarationStmt() = 0;
+    virtual void codegen(Code &) = 0;
+};
+
+struct VariableDeclarationStmt : DeclarationStmt
+{
+    String name;
     Ptr<Expr> expr;
     bool is_ref;
 
-    VariableDeclarationStmt(
-        Ptr<IdentifierExpr> &&id,
-        Ptr<Expr> &&expr,
-        bool is_ref = false)
-      : id(std::move(id))
+    VariableDeclarationStmt(String name,
+        Ptr<Expr> &&expr, bool is_ref = false)
+      : name(std::move(name))
       , expr(std::move(expr))
       , is_ref(is_ref)
     {
@@ -360,7 +380,7 @@ struct VariableDeclarationStmt : Stmt
     void codegen(Code &) override;
 };
 
-struct MultiVarsDeclarationStmt : Stmt
+struct MultiVarsDeclarationStmt : DeclarationStmt
 {
     // just reuse VariableDeclarationStmt
     std::list<VariableDeclarationStmt> decls;
@@ -378,29 +398,12 @@ struct MultiVarsDeclarationStmt : Stmt
     void codegen(Code &) override;
 };
 
-struct FunctionDeclarationStmt : Stmt
-{
-    Ptr<IdentifierExpr> id;
-    Ptr<LambdaExpr> lambda;
-
-    FunctionDeclarationStmt(
-        Ptr<IdentifierExpr> &&id,
-        Ptr<LambdaExpr> &&lambda)
-      : id(std::move(id))
-      , lambda(std::move(lambda))
-    {
-        // ...
-    }
-
-    void codegen(Code &) override;
-};
-
 struct PrefixopDeclarationStmt : Stmt
 {
-    Ptr<IdentifierExpr> id;
+    std::string op;
 
-    PrefixopDeclarationStmt(Ptr<IdentifierExpr> &&id)
-      : id(std::move(id))
+    PrefixopDeclarationStmt(std::string op)
+      : op(std::move(op))
     {
         // ...
     }
@@ -411,11 +414,10 @@ struct PrefixopDeclarationStmt : Stmt
 struct InfixopDeclarationStmt : Stmt
 {
     Size priority;
-    Ptr<IdentifierExpr> id;
+    std::string op;
 
-    InfixopDeclarationStmt(Ptr<IdentifierExpr> &&id,
-        Size priority)
-      : priority(priority), id(std::move(id))
+    InfixopDeclarationStmt(std::string op, Size priority)
+      : priority(priority), op(std::move(op))
     {
         // ...
     }
@@ -500,14 +502,14 @@ struct DoWhileStmt : Stmt
 struct ForeachStmt : Stmt
 {
     Ptr<Expr> expr;
-    Ptr<IdentifierExpr> id;
+    std::string varname;
     Ptr<BlockExpr> block;
 
     ForeachStmt(Ptr<Expr> &&expr,
-        Ptr<IdentifierExpr> &&id,
+        std::string varname,
         Ptr<BlockExpr> &&block)
       : expr(std::move(expr))
-      , id(std::move(id))
+      , varname(std::move(varname))
       , block(std::move(block))
     {
         // ...