+++ src/compiler/code.cpp
@@ -325,6 +325,9 @@ void Code::print(ostream &out)
         case Opcode::NewScope:
             printer.add_line(i, "NewScope");
             break;
+        case Opcode::EndScope:
+            printer.add_line(i, "EndScope");
+            break;
 
         case Opcode::CallAc:
             printer.add_line(i, "CallAc");
@@ -455,6 +458,9 @@ void Code::print(ostream &out)
         case Opcode::BuildDict:
             printer.add_line(i, "BuildDict", OPRAND(Size));
             break;
+        case Opcode::BuildClass:
+            printer.add_line(i, "BuildClass", OPRAND(String));
+            break;
         }
     }
     printer.print();
@@ -524,6 +530,7 @@ void Code::serialize(ostream &out)
         case Opcode::StoreRef:
         case Opcode::StoreLocal:
         case Opcode::AddPrefixOp:
+        case Opcode::BuildClass:
             typeout(out, OPRAND(String));
             break;
 
@@ -643,6 +650,7 @@ bool Code::unserialize(ifstream &in)
         case Opcode::StoreRef:
         case Opcode::StoreLocal:
         case Opcode::AddPrefixOp:
+        case Opcode::BuildClass:
         {
             String val;
             typein(in, val);