+++ src/objects/classobject.hpp
@@ -6,23 +6,22 @@ namespace anole
 {
 class Scope;
 
-class EnumObject : public Object
+class ClassObject : public Object
 {
   public:
-    EnumObject(SPtr<Scope> scope)
-      : Object(ObjectType::Enum)
-      , scope_(std::move(scope))
-    {
-        // ...
-    }
+    ClassObject(String name,
+        SPtr<Scope> pre_scope
+    );
 
     Address load_member(const String &name) override;
+    void call(Size num) override;
 
-    void collect(std::function<void(Scope *)>) override;
+    // collect
 
     SPtr<Scope> &scope() { return scope_; }
 
   private:
+    String name_;
     SPtr<Scope> scope_;
 };
-}
+} // namespace anole