+++ src/objects/funcobject.cpp
@@ -13,7 +13,7 @@ FunctionObject::FunctionObject(SPtr<Scope> pre_scope,
     SPtr<Code> code, Size base,
     Size parameter_num)
   : Object(ObjectType::Func)
-  , scope_(std::make_shared<Scope>(pre_scope))
+  , scope_(make_shared<Scope>(pre_scope))
   , code_(code), base_(base)
   , parameter_num_(parameter_num)
 {
@@ -133,6 +133,11 @@ void FunctionObject::call(Size num)
             throw RuntimeError("missing the parameter named '" + OPRAND(String) + '\'');
         }
     }
+}
+
+bool FunctionObject::is_callable()
+{
+    return true;
 }
 
 void FunctionObject::collect(function<void(Scope *)> func)