+++ src/objects/object.hpp
@@ -34,7 +34,10 @@ enum ObjectType : int
     Thunk,
     Cont,
     AnoleModule,
-    CppModule
+    CppModule,
+    Class,
+    Method,
+    Instance,
 };
 }
 using object_type::ObjectType;
@@ -84,6 +87,7 @@ class Object
     virtual Address load_member(const String &name);
 
     virtual void call(Size num);
+    virtual bool is_callable();
 
     virtual void collect(std::function<void(Scope *)>);
     virtual void collect(std::function<void(Object *)>);